requirements.txt
absl-py==0.4.0
alabaster==0.7.11
apturl==0.5.2
astor==0.7.1
astroid==2.0.4
Babel==2.6.0
backcall==0.1.0
beautifulsoup4==4.4.1
bleach==2.1.4
blinker==1.3
Brlapi==0.6.4
chardet==2.3.0
checkbox-support==0.22
cloudpickle==0.5.5
command-not-found==0.3
cryptography==1.2.3
cycler==0.10.0
Cython==0.28.2
decorator==4.3.0
defer==1.0.6
docutils==0.14
entrypoints==0.2.3
feedparser==5.1.3
future==0.16.0
gast==0.2.0
graphviz==0.8.3
grpcio==1.14.1
guacamole==0.9.2
h5py==2.7.1
html5lib==1.0.1
httplib2==0.9.1
hyperopt==0.1
idna==2.0
imagesize==1.0.0
ipykernel==4.8.2
ipython==6.5.0
ipython-genutils==0.2.0
isort==4.3.4
jedi==0.12.1
jeepney==0.3.1
Jinja2==2.8
jsonschema==2.6.0
jupyter-client==5.2.3
jupyter-core==4.4.0
keyring==13.2.1
kiwisolver==1.0.1
language-selector==0.1
lazy-object-proxy==1.3.1
louis==2.6.4
lxml==3.5.0
Mako==1.0.3
Markdown==2.6.11
MarkupSafe==0.23
matplotlib==2.2.2
mccabe==0.6.1
mistune==0.8.3
nbconvert==5.3.1
nbformat==4.4.0
networkx==1.11
nose==1.3.7
numpy==1.14.5
numpydoc==0.8.0
oauthlib==1.0.3
onboard==1.2.0
opencv-python==3.4.3.18
packaging==17.1
padme==1.1.1
pandas==0.23.0
pandocfilters==1.4.2
parso==0.3.1
pexpect==4.0.1
pickleshare==0.7.4
Pillow==3.1.2
pkg-resources==0.0.0
plainbox==0.25
prompt-toolkit==1.0.15
protobuf==3.6.1
psutil==5.4.7
ptyprocess==0.5
pyasn1==0.1.9
pycodestyle==2.4.0
pycups==1.9.73
pycurl==7.43.0
pydot==1.2.4
pyflakes==2.0.0
Pygments==2.2.0
pygobject==3.20.0
PyJWT==1.3.0
pylint==2.1.1
pymongo==3.7.1
pyparsing==2.2.0
PyQt5==5.9.2
python-apt==1.1.0b1
python-dateutil==2.5.0
python-debian==0.1.27
python-systemd==231
pytz==2014.10
PyWavelets==0.5.2
pyxdg==0.25
pyzmq==17.1.2
QtAwesome==0.4.4
qtconsole==4.4.1
QtPy==1.5.0
reportlab==3.3.0
requests==2.9.1
rope==0.11.0
rope-py3k==0.9.4.post1
scikit-image==0.13.1
scikit-learn==0.19.2
scipy==1.1.0
screen-resolution-extra==0.0.0
seaborn==0.8.1
SecretStorage==3.0.1
sessioninstaller==0.0.0
simplegeneric==0.8.1
sip==4.19.8
six==1.10.0
snowballstemmer==1.2.1
Sphinx==1.7.7
sphinxcontrib-websupport==1.1.0
spyder==3.3.1
spyder-kernels==0.2.6
ssh-import-id==5.5
system-service==0.3
tensorboard==1.10.0
tensorflow-gpu==1.10.1
termcolor==1.1.0
testpath==0.3.1
Theano==1.0.1
tornado==5.1
traitlets==4.3.2
typed-ast==1.1.0
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
unity-scope-calculator==0.1
unity-scope-chromiumbookmarks==0.1
unity-scope-colourlovers==0.1
unity-scope-devhelp==0.1
unity-scope-firefoxbookmarks==0.1
unity-scope-gdrive==0.7
unity-scope-manpages==0.1
unity-scope-openclipart==0.1
unity-scope-texdoc==0.1
unity-scope-tomboy==0.1
unity-scope-virtualbox==0.1
unity-scope-yelp==0.1
unity-scope-zotero==0.1
urllib3==1.13.1
usb-creator==0.3.0
virtualenv==15.0.1
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
wrapt==1.10.11
xdiagnose==3.8.4.1
xkit==0.0.0
XlsxWriter==0.7.3
